Se sube el proyecto

+++ modelos/paqueteModelo.php
@@ -20,7 +20,7 @@ class paqueteModelo
 
         $this->db = new mysqli($host, $user, $pass, $db);
         mysqli_set_charset($this->db, 'utf8');
-        
+
     }
     public function obtenerPaquetePorCodigo($codigo)
     {
@@ -34,30 +34,14 @@ class paqueteModelo
     }
     public function obtenerPaquete($id_paquete)
     {
-        $paquete1 = [];
-        $instruccion = "SELECT * FROM paquete INNER JOIN almacena ON almacena.id_paquete = paquete.id_paquete INNER JOIN recoge ON recoge.id_almacen_cliente = almacena.id_almacen_cliente WHERE paquete.id_paquete='$id_paquete'";
-        $resultado = mysqli_query($this->db, $instruccion);    
+
+        $paquete = [];
+        $instruccion = "SELECT *, paquete.estado AS paquete_estado FROM paquete INNER JOIN destino ON paquete.id_destino = destino.id_destino INNER JOIN almacena ON paquete.id_paquete = almacena.id_paquete INNER JOIN tiene ON almacena.id_almacen_cliente = tiene.id_almacen_cliente INNER JOIN empresa_cliente ON tiene.id_empresa_cliente = empresa_cliente.id_empresa_cliente WHERE paquete.id_paquete='$id_paquete'";
+        $resultado = mysqli_query($this->db, $instruccion);
         while ($row = mysqli_fetch_assoc($resultado)) {
-            array_push($paquete1, $row);
+            array_push($paquete, $row);
         }
-
-        if (count($paquete1) > 0){
-            $paquete = [];
-            $instruccion = "SELECT *, paquete.estado AS paquete_estado FROM paquete INNER JOIN destino ON paquete.id_destino = destino.id_destino INNER JOIN almacena ON paquete.id_paquete = almacena.id_paquete INNER JOIN tiene ON almacena.id_almacen_cliente = tiene.id_almacen_cliente INNER JOIN empresa_cliente ON tiene.id_empresa_cliente = empresa_cliente.id_empresa_cliente INNER JOIN recoge ON almacena.id_almacen_cliente = recoge.id_almacen_cliente INNER JOIN camioneta ON recoge.id_camioneta = camioneta.id_camioneta INNER JOIN vehiculo ON camioneta.id_camioneta = vehiculo.id_vehiculo WHERE paquete.id_paquete='$id_paquete' AND recoge.fecha_recogida IS NULL ORDER BY fecha_recogida_ideal DESC";
-            $resultado = mysqli_query($this->db, $instruccion);    
-            while ($row = mysqli_fetch_assoc($resultado)) {
-                array_push($paquete, $row);
-            }
-        } else {
-            $paquete = [];
-            $instruccion = "SELECT *, paquete.estado AS paquete_estado FROM paquete INNER JOIN destino ON paquete.id_destino = destino.id_destino INNER JOIN almacena ON paquete.id_paquete = almacena.id_paquete INNER JOIN tiene ON almacena.id_almacen_cliente = tiene.id_almacen_cliente INNER JOIN empresa_cliente ON tiene.id_empresa_cliente = empresa_cliente.id_empresa_cliente WHERE paquete.id_paquete='$id_paquete'";
-            $resultado = mysqli_query($this->db, $instruccion);    
-            while ($row = mysqli_fetch_assoc($resultado)) {
-                array_push($paquete, $row);
-            }
-        }
-
-         return $paquete;
+        return $paquete;
     }
     public function obtenerPaquetePorEmpresa($empresa)
     {
@@ -89,36 +73,36 @@ class paqueteModelo
         $sentencia_almacen = "AND almacena.id_almacen_cliente=$id_almacen_cliente ";
         $sentencia_estado = "AND paquete.estado='$estado' ";
 
-        if(!isset($id_almacen_cliente) || is_null($id_almacen_cliente) || empty(trim($id_almacen_cliente))){
+        if (!isset($id_almacen_cliente) || is_null($id_almacen_cliente) || empty(trim($id_almacen_cliente))) {
             $sentencia_almacen = "";
-            if(!isset($id_paquete) || is_null($id_paquete) || empty(trim($id_paquete))){
+            if (!isset($id_paquete) || is_null($id_paquete) || empty(trim($id_paquete))) {
                 $sentencia_paquete = "";
-                if(!isset($estado) || is_null($estado) || empty(trim($estado))){
+                if (!isset($estado) || is_null($estado) || empty(trim($estado))) {
                     $sentencia_estado = "";
                 } else {
                     $sentencia_estado = "WHERE estado='$estado' ";
                 }
-            } else{
-                if(!isset($estado) || is_null($estado) || empty(trim($estado))){
+            } else {
+                if (!isset($estado) || is_null($estado) || empty(trim($estado))) {
                     $sentencia_paquete = "WHERE paquete.id_paquete=$id_paquete ";
                     $sentencia_estado = "";
-                } else{
+                } else {
                     $sentencia_paquete = "WHERE paquete.id_paquete=$id_paquete ";
                     $sentencia_estado = "AND paquete.estado='$estado' ";
                 }
             }
-        } else{
-            if(!isset($id_paquete) || is_null($id_paquete) || empty(trim($id_paquete))){
+        } else {
+            if (!isset($id_paquete) || is_null($id_paquete) || empty(trim($id_paquete))) {
                 $sentencia_almacen = "WHERE almacena.id_almacen_cliente=$id_almacen_cliente ";
                 $sentencia_paquete = "";
             }
-            if(!isset($estado) || is_null($estado) || empty(trim($estado))){
+            if (!isset($estado) || is_null($estado) || empty(trim($estado))) {
                 $sentencia_estado = "";
             }
- 
+
         }
         $paquete = [];
-        $instruccion = "SELECT * FROM paquete INNER JOIN almacena ON paquete.id_paquete = almacena.id_paquete " . $sentencia_paquete . $sentencia_almacen . $sentencia_estado . "ORDER BY paquete.id_paquete ASC";
+        $instruccion = "SELECT * FROM paquete INNER JOIN almacena ON paquete.id_paquete = almacena.id_paquete " . $sentencia_paquete . $sentencia_almacen . $sentencia_estado . " INNER JOIN destino ON paquete.id_destino = destino.id_destino ORDER BY paquete.id_paquete ASC";
         $resultado = mysqli_query($this->db, $instruccion);
         while ($row = mysqli_fetch_assoc($resultado)) {
             array_push($paquete, $row);
@@ -131,7 +115,7 @@ class paqueteModelo
     {
         $where = ($id_paquete == null) ? "" : " WHERE id_paquete='$id_paquete'";
         $paquetes = [];
-        $instruccion = "SELECT * FROM paquete INNER JOIN destino ON paquete.id_destino = destino.id_destino" . $where . " ORDER BY paquete.id_paquete ASC" ;
+        $instruccion = "SELECT * FROM paquete INNER JOIN destino ON paquete.id_destino = destino.id_destino" . $where . " ORDER BY paquete.id_paquete ASC";
         $resultado = mysqli_query($this->db, $instruccion);
         while ($row = mysqli_fetch_assoc($resultado)) {
             array_push($paquetes, $row);
@@ -166,41 +150,41 @@ class paqueteModelo
             'respuesta' => "No existe el paquete con la ID " . $id_paquete
         ];
         if (count($validar) > 0) {
-            if ($tipo_usu !== "empresa"){
-            $instruccion = "UPDATE paquete SET direccion='$direccion', peso='$peso', volumen='$volumen', fragil='$fragil', estado='$estado' WHERE id_paquete='$id_paquete'";
-            mysqli_query($this->db, $instruccion);
-            $resultado = [
-                'error' => "Éxito",
-                'respuesta' => "Paquete modificado"
-            ];
+            if ($tipo_usu !== "empresa") {
+                $instruccion = "UPDATE paquete SET direccion='$direccion', peso='$peso', volumen='$volumen', fragil='$fragil', estado='$estado' WHERE id_paquete='$id_paquete'";
+                mysqli_query($this->db, $instruccion);
+                $resultado = [
+                    'error' => "Éxito",
+                    'respuesta' => "Paquete modificado"
+                ];
             } else {
                 $instruccion = "SELECT empresa_cliente.nombre_de_empresa, paquete.estado FROM paquete INNER JOIN destino ON paquete.id_destino = destino.id_destino INNER JOIN almacena ON paquete.id_paquete = almacena.id_paquete INNER JOIN tiene ON almacena.id_almacen_cliente = tiene.id_almacen_cliente INNER JOIN empresa_cliente ON tiene.id_empresa_cliente = empresa_cliente.id_empresa_cliente WHERE paquete.id_paquete='$id_paquete'";
                 $resultado = mysqli_query($this->db, $instruccion);
-                $fila =  mysqli_fetch_assoc($resultado);
+                $fila = mysqli_fetch_assoc($resultado);
                 $empresa1 = $fila["nombre_de_empresa"];
                 $estado = $fila["estado"];
-                if ($estado == "En almacén cliente"){
-                    if ($empresa == $empresa1){
-                    $instruccion = "UPDATE paquete SET direccion='$direccion', peso='$peso', volumen='$volumen', fragil='$fragil' WHERE id_paquete='$id_paquete'";
-                    mysqli_query($this->db, $instruccion);
-                    $resultado = [
-                        'error' => "Éxito",
-                        'respuesta' => "Paquete modificado"
-                    ];
+                if ($estado == "En almacén cliente") {
+                    if ($empresa == $empresa1) {
+                        $instruccion = "UPDATE paquete SET direccion='$direccion', peso='$peso', volumen='$volumen', fragil='$fragil' WHERE id_paquete='$id_paquete'";
+                        mysqli_query($this->db, $instruccion);
+                        $resultado = [
+                            'error' => "Éxito",
+                            'respuesta' => "Paquete modificado"
+                        ];
                     } else {
-                    $resultado = [
-                        'error' => "Error",
-                        'respuesta' => "Este paquete no es de tu pertenencia"
-                    ];
+                        $resultado = [
+                            'error' => "Error",
+                            'respuesta' => "Este paquete no es de tu pertenencia"
+                        ];
                     }
-                } else{
+                } else {
                     $resultado = [
                         'error' => "Error",
                         'respuesta' => "No tienes permiso para hacer esto"
                     ];
                 }
 
- 
+
             }
 
         }
@@ -217,10 +201,10 @@ class paqueteModelo
         if (count($validar) > 0) {
             $instruccion = "SELECT * FROM paquete WHERE id_paquete='$id_paquete'";
             $resultado = mysqli_query($this->db, $instruccion);
-            $fila =  mysqli_fetch_assoc($resultado);
+            $fila = mysqli_fetch_assoc($resultado);
             $estado = $fila["estado"];
-            if ($estado == "En almacén cliente" || $estado == "En camioneta (central)"){
-                if ($estado == "En almacén cliente"){
+            if ($estado == "En almacén cliente" || $estado == "En camioneta (central)") {
+                if ($estado == "En almacén cliente") {
                     $instruccion = "UPDATE paquete SET estado='En camioneta (central)' WHERE id_paquete = '$id_paquete'";
                     mysqli_query($this->db, $instruccion);
 
@@ -257,8 +241,8 @@ class paqueteModelo
                     </body>
                     </html>
                     ';
-                    $mail->send();   
-                    
+                    $mail->send();
+
                     $resultado = [
                         'error' => "Éxito",
                         'respuesta' => "Paquete recogido"
@@ -269,13 +253,13 @@ class paqueteModelo
                     $resultado = [
                         'error' => "Éxito",
                         'respuesta' => "Paquete desrecogido"
-                    ];          
+                    ];
                 }
             } else {
                 $resultado = [
                     'error' => "Error",
                     'respuesta' => "No tienes permisos para hacer esto"
-                ]; 
+                ];
             }
         }
         return $resultado;
@@ -289,24 +273,62 @@ class paqueteModelo
             'respuesta' => "No existe el paquete con la ID " . $id_paquete
         ];
         if (count($validar) > 0) {
-            if ($tipo_usu !== "empresa"){
-                $instruccion = "DELETE FROM almacena WHERE id_paquete='$id_paquete'";
-                mysqli_query($this->db, $instruccion);
+            if ($tipo_usu !== "empresa") {
+                $instruccion = "SELECT * FROM paquete WHERE id_paquete='$id_paquete'";
+                $resultado = mysqli_query($this->db, $instruccion);
+                $fila = mysqli_fetch_assoc($resultado);
+                $estado = $fila["estado"];
 
-                $instruccion = "DELETE FROM paquete WHERE id_paquete='$id_paquete'";
-                mysqli_query($this->db, $instruccion);
-                $resultado = [
-                'error' => "Éxito",
-                'respuesta' => "Paquete eliminado"
-            ];
+                if ($estado != "En almacén central (lote)") {
+
+                    if ($estado != "En camión (plataforma)"){
+
+                        if ($estado != "En camioneta (central)"){
+
+                            if ($estado !== "Entregado") {
+
+                            $instruccion = "DELETE FROM almacena WHERE id_paquete='$id_paquete'";
+                            mysqli_query($this->db, $instruccion);
+        
+                            $instruccion = "DELETE FROM paquete WHERE id_paquete='$id_paquete'";
+                            mysqli_query($this->db, $instruccion);
+                            $resultado = [
+                                'error' => "Éxito",
+                                'respuesta' => "Paquete eliminado"
+                            ];
+                        } else {
+                            $resultado = [
+                                'error' => "Error",
+                                'respuesta' => "Hay paquetes entregados"
+                            ];
+                        }
+                        } else{
+                            $resultado = [
+                                'error' => "Error",
+                                'respuesta' => "Hay paquetes en camionetas"
+                            ];
+                        }
+                    } else {
+                        $resultado = [
+                            'error' => "Error",
+                            'respuesta' => "Hay paquetes en camiones"
+                        ];
+                    }
+
+                } else {
+                    $resultado = [
+                        'error' => "Error",
+                        'respuesta' => "Hay paquetes en lotes"
+                    ];
+                }
             } else {
                 $instruccion = "SELECT empresa_cliente.nombre_de_empresa, paquete.estado FROM paquete INNER JOIN destino ON paquete.id_destino = destino.id_destino INNER JOIN almacena ON paquete.id_paquete = almacena.id_paquete INNER JOIN tiene ON almacena.id_almacen_cliente = tiene.id_almacen_cliente INNER JOIN empresa_cliente ON tiene.id_empresa_cliente = empresa_cliente.id_empresa_cliente WHERE paquete.id_paquete='$id_paquete'";
                 $resultado = mysqli_query($this->db, $instruccion);
                 $fila = mysqli_fetch_assoc($resultado);
                 $empresa1 = $fila["nombre_de_empresa"];
                 $estado = $fila["estado"];
-                if ($estado == "En almacén cliente"){
-                    if ($empresa == $empresa1){
+                if ($estado == "En almacén cliente") {
+                    if ($empresa == $empresa1) {
                         $instruccion = "DELETE FROM almacena WHERE id_paquete='$id_paquete'";
                         mysqli_query($this->db, $instruccion);
 
@@ -322,14 +344,14 @@ class paqueteModelo
                             'respuesta' => "Este paquete no es de tu pertenencia"
                         ];
                     }
-                } else{
+                } else {
                     $resultado = [
                         'error' => "Error",
                         'respuesta' => "No tienes permiso para hacer esto"
                     ];
-                }  
-                    
- 
+                }
+
+
             }
 
         }
